V2 - Move V2 files into root folder, move V1 into own folder
