+++ SMLHelper/Assets/ModSprite.cs
