+++ SMLHelper/Handlers/CrossModSupportHandler.cs
