+++ SMLHelper/Handlers/SpriteHandler.cs
