+++ SMLHelper/Legacy/CraftNodeToScrub.cs
