+++ SMLHelper/Legacy/CraftScheme.cs
