+++ SMLHelper/Legacy/CustomPrefabHandler.cs
