+++ SMLHelper/Legacy/CustomSpriteHandler.cs
