+++ SMLHelper/Patchers/CraftTreeTypePatcher.cs
@@ -1,24 +1,143 @@
-﻿using CraftTreeTypePatcher2 = SMLHelper.V2.Patchers.CraftTreeTypePatcher;
-using CraftTreePatcher2 = SMLHelper.V2.Patchers.CraftTreePatcher;
-using CustomCraftTreeRoot2 = SMLHelper.V2.Crafting.CustomCraftTreeRoot;
-
-namespace SMLHelper.Patchers
+﻿namespace SMLHelper.V2.Patchers
 {
-    [System.Obsolete("Use SMLHelper.V2 instead.")]
-    public class CraftTreeTypePatcher
+    using Harmony;
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Reflection;
+    using Utilities;
+
+    internal class CraftTreeTypePatcher
     {
-        [System.Obsolete("Use SMLHelper.V2 instead.")]
-        public static CustomCraftTreeRoot CreateCustomCraftTreeAndType(string name, out CraftTree.Type craftTreeType)
+        #region Internal Fields
+
+        internal const int startingIndex = 11; // The default CraftTree.Type contains indexes 0 through 10
+        internal static readonly EnumCacheManager<CraftTree.Type> cacheManager = new EnumCacheManager<CraftTree.Type>("CraftTreeType", startingIndex);
+
+        #endregion
+
+        #region Adding CraftTreeTypes and TreeRoots
+
+        internal static Crafting.CustomCraftTreeRoot CreateCustomCraftTreeAndType(string name, out CraftTree.Type craftTreeType)
         {
-            var customCraftTreeRoot2 = CraftTreeTypePatcher2.CreateCustomCraftTreeAndType(name, out craftTreeType);
-            CraftTreePatcher2.CustomTrees.Remove(craftTreeType);
+            var cache = cacheManager.GetCacheForTypeName(name);
 
-            var customCraftTreeRoot = new CustomCraftTreeRoot(craftTreeType, name);
+            if (cache == null)
+            {
+                cache = new EnumTypeCache()
+                {
+                    Name = name,
+                    Index = cacheManager.GetNextFreeIndex()
+                };
+            }
 
-            CraftTreePatcher.CustomTrees.Add(craftTreeType, customCraftTreeRoot);
+            if (cacheManager.IsIndexConflicting(cache.Index))
+                cache.Index = cacheManager.GetNextFreeIndex();
 
-            return customCraftTreeRoot;
+            craftTreeType = (CraftTree.Type)cache.Index;
+
+            cacheManager.customEnumTypes.Add(craftTreeType, cache);
+
+            Logger.Log("Successfully added CraftTree Type: \"{0}\" to Index: \"{1}\"", name, cache.Index);
+
+            cacheManager.SaveCache();
+
+            var customTreeRoot = new Crafting.CustomCraftTreeRoot(craftTreeType, name);
+
+            CraftTreePatcher.CustomTrees[craftTreeType] = customTreeRoot;
+
+            return customTreeRoot;
         }
 
+        #endregion
+
+        #region Patches
+
+        internal static void Patch(HarmonyInstance harmony)
+        {
+            var enumType = typeof(Enum);
+            var thisType = typeof(CraftTreeTypePatcher);
+            var techTypeType = typeof(CraftTree.Type);
+
+            harmony.Patch(enumType.GetMethod("GetValues", BindingFlags.Public | BindingFlags.Static), null,
+                new HarmonyMethod(thisType.GetMethod("Postfix_GetValues", BindingFlags.NonPublic | BindingFlags.Static)));
+
+            harmony.Patch(enumType.GetMethod("IsDefined", BindingFlags.Public | BindingFlags.Static),
+                new HarmonyMethod(thisType.GetMethod("Prefix_IsDefined", BindingFlags.NonPublic | BindingFlags.Static)), null);
+
+            harmony.Patch(enumType.GetMethod("Parse", new Type[] { typeof(Type), typeof(string), typeof(bool) }),
+                new HarmonyMethod(thisType.GetMethod("Prefix_Parse", BindingFlags.NonPublic | BindingFlags.Static)), null);
+
+            harmony.Patch(techTypeType.GetMethod("ToString", new Type[0]),
+                new HarmonyMethod(thisType.GetMethod("Prefix_ToString", BindingFlags.NonPublic | BindingFlags.Static)), null);
+
+            Logger.Log("CraftTreeTypePatcher is done.");
+        }
+
+        internal static void Postfix_GetValues(Type enumType, ref Array __result)
+        {
+            if (enumType.Equals(typeof(CraftTree.Type)))
+            {
+                var listArray = new List<CraftTree.Type>();
+                foreach (var obj in __result)
+                {
+                    listArray.Add((CraftTree.Type)obj);
+                }
+
+                __result = listArray
+                    .Concat(cacheManager.customEnumTypes.Keys)
+                    .ToArray();
+            }
+        }
+
+        internal static bool Prefix_IsDefined(Type enumType, object value, ref bool __result)
+        {
+            if (enumType.Equals(typeof(CraftTree.Type)))
+            {
+                if (cacheManager.customEnumTypes.Keys.Contains((CraftTree.Type)value))
+                {
+                    __result = true;
+                    return false;
+                }
+            }
+
+            return true;
+        }
+
+        internal static bool Prefix_Parse(Type enumType, string value, bool ignoreCase, ref object __result)
+        {
+            if (enumType.Equals(typeof(CraftTree.Type)))
+            {
+                foreach (var techType in cacheManager.customEnumTypes)
+                {
+                    if (value.Equals(techType.Value.Name, ignoreCase ? StringComparison.InvariantCultureIgnoreCase : StringComparison.InvariantCulture))
+                    {
+                        __result = techType.Key;
+                        return false;
+                    }
+                }
+            }
+
+            return true;
+        }
+
+        internal static bool Prefix_ToString(Enum __instance, ref string __result)
+        {
+            if (__instance.GetType().Equals(typeof(CraftTree.Type)))
+            {
+                foreach (var techType in cacheManager.customEnumTypes)
+                {
+                    if (__instance.Equals(techType.Key))
+                    {
+                        __result = techType.Value.Name;
+                        return false;
+                    }
+                }
+            }
+
+            return true;
+        }
+
+        #endregion
     }
 }
