+++ SMLHelper/Patchers/DevConsolePatcher.cs
@@ -1,41 +1,61 @@
-﻿using System.Reflection;
-using System.Collections.Generic;
-using DevConsolePatcher2 = SMLHelper.V2.Patchers.DevConsolePatcher;
-using CommandInfo2 = SMLHelper.V2.Patchers.CommandInfo;
-
-namespace SMLHelper.Patchers
+﻿namespace SMLHelper.V2.Patchers
 {
-    [System.Obsolete("Use SMLHelper.V2 instead.")]
-    public class DevConsolePatcher
+    using Harmony;
+    using System;
+    using System.Collections.Generic;
+    using System.Linq;
+    using System.Reflection;
+
+    internal class DevConsolePatcher
     {
-        [System.Obsolete("Use SMLHelper.V2 instead.")]
         public static List<CommandInfo> commands = new List<CommandInfo>();
 
-        internal static void Patch()
+        public static void Patch(HarmonyInstance harmony)
         {
-            commands.ForEach(x => DevConsolePatcher2.commands.Add(x.GetV2CommandInfo()));
+            var devConsoleType = typeof(DevConsole);
+            var thisType = typeof(DevConsolePatcher);
+            var submitMethod = devConsoleType.GetMethod("Submit", BindingFlags.Instance | BindingFlags.NonPublic);
 
-            V2.Logger.Log("Old DevConsolePatcher is done.");
+            harmony.Patch(submitMethod, null, new HarmonyMethod(thisType.GetMethod("Postfix", BindingFlags.Static | BindingFlags.NonPublic)));
+            Logger.Log("DevConsolePatcher is done.");
+        }
+
+        internal static void Postfix(bool __result, string value)
+        {
+            var separator = new char[]
+            {
+                ' ',
+                '\t'
+            };
+
+            var text = value.Trim();
+            var args = text.Split(separator, StringSplitOptions.RemoveEmptyEntries);
+
+            if (args.Length != 0)
+            {
+                foreach (var command in commands)
+                {
+                    if (command.Name.Contains(args[0]))
+                    {
+                        var argsList = args.ToList();
+                        argsList.RemoveAt(0);
+                        var newArgs = argsList.ToArray();
+                        command.CommandHandler.Invoke(null, new object[] { newArgs });
+                        __result = true;
+                        return;
+                    }
+                }
+            }
+
+            __result = false;
         }
     }
 
-    [System.Obsolete("Use SMLHelper.V2 instead.")]
-    public class CommandInfo
+    internal class CommandInfo
     {
         public MethodInfo CommandHandler;
         public string Name;
         public bool CaseSensitive;
         public bool CombineArgs;
-
-        internal CommandInfo2 GetV2CommandInfo()
-        {
-            var commandInfo = new CommandInfo2();
-            commandInfo.CommandHandler = CommandHandler;
-            commandInfo.Name = Name;
-            commandInfo.CaseSensitive = CaseSensitive;
-            commandInfo.CombineArgs = CombineArgs;
-
-            return commandInfo;
-        }
     }
 }
