+++ SMLHelper/Patchers/PDAEncyclopediaPatcher.cs
