+++ SMLHelper/Patchers/PrefabDatabasePatcher.cs
