+++ SMLHelper/Patchers/ResourcesPatcher.cs
