+++ SMLHelper/Patchers/SpritePatcher.cs
