+++ SMLHelper/Util/EnumCacheManager.cs
