+++ SMLHelper/Utility.cs
@@ -1,25 +1,13 @@
-﻿#pragma warning disable CS0618 // Type or member is obsolete
-using System;
-using System.Collections;
-using System.Reflection;
-using System.IO;
-using UnityEngine;
-using System.Collections.Generic;
-
-namespace SMLHelper
+﻿namespace SMLHelper.V2
 {
+    using System;
+    using System.Collections;
+    using System.Reflection;
+    using System.IO;
+    using UnityEngine;
+
     public class Utility
     {
-        public readonly static Dictionary<CraftScheme, CraftTree.Type> CraftSchemeMap = new Dictionary<CraftScheme, CraftTree.Type>
-        {
-            { CraftScheme.Constructor, CraftTree.Type.Constructor },
-            { CraftScheme.CyclopsFabricator, CraftTree.Type.CyclopsFabricator },
-            { CraftScheme.Fabricator, CraftTree.Type.Fabricator },
-            { CraftScheme.MapRoom, CraftTree.Type.MapRoom },
-            { CraftScheme.SeamothUpgrades, CraftTree.Type.SeamothUpgrades },
-            { CraftScheme.Workbench, CraftTree.Type.Workbench },
-        };
-
         public static void AddBasicComponents(ref GameObject _object, string classId)
         {
             var rb = _object.AddComponent<Rigidbody>();
@@ -47,61 +35,11 @@ namespace SMLHelper
         public static void PatchDictionary(Type type, string name, IDictionary dictionary, BindingFlags flags)
         {
             var dictionaryField = type.GetField(name, flags);
-            var craftDataDict = dictionaryField.GetValue(null) as IDictionary;
+            var dict = dictionaryField.GetValue(null) as IDictionary;
 
-            PatchDictionaryInternal(craftDataDict, dictionary);
-        }
-
-        private static void PatchDictionaryInternal(IDictionary origDict, IDictionary additions)
-        {
-            foreach (DictionaryEntry entry in additions)
+            foreach(DictionaryEntry entry in dictionary)
             {
-                //As long as we have the type we wanna check, and the instance - "is" would do the trick
-                if (entry.Value is IDictionary)
-                {
-                    //You have to check the KEYS for being equal. Not just their type, since TechType.Titanium and TechType.AcidMushroom has equal types.
-                    foreach (DictionaryEntry origDictEntry in origDict)
-                    {
-                        if (origDictEntry.Key.Equals(entry.Key))
-                        {
-                            PatchDictionaryInternal((IDictionary)origDictEntry.Value, (IDictionary)entry.Value);
-                        }
-                    }
-                }
-
-                if (entry.Value is IList)
-                {
-                    foreach (DictionaryEntry origDictEntry in origDict)
-                    {
-                        //Ok, see - here is the problem. You were updating original list. Right?
-                        //Adding a new value to it.
-                        if (origDictEntry.Key.Equals(entry.Key))
-                        {
-                            //Now - we are modifying the entry list, so it would contain all the default data.
-                            //There should be over way, but it would be way overcoded, and I didn't come to a better solution....
-                            PatchListInternal((IList)entry.Value, (IList)origDictEntry.Value);
-                        }
-                    }
-                }
-                //But here - you were setting the value of the original dict to UNMODIFIED entry list.
-                //While the original value already modified....
-                origDict[entry.Key] = entry.Value;
-            }
-        }
-
-        private static void PatchListInternal(IList orig, IList additions)
-        {
-            foreach (var addition in additions)
-            {
-                if (addition is IList)
-                {
-                    foreach (var origEntry in orig)
-                    {
-                        if (origEntry.GetType() == addition.GetType())
-                            PatchListInternal((IList)origEntry, (IList)addition);
-                    }
-                }
-                orig.Add(addition);
+                dict.Add(entry.Key, entry.Value);
             }
         }
